requirements.txt
pandas==2.2.2
numpy==1.26.4
PyYAML==6.0.2
yfinance==0.2.54
feedparser==6.0.11
requests==2.32.3
scipy==1.13.1
pyarrow==17.0.0
